requirements.txt
Flask==3.0.2
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
python-dotenv==1.1.0
psycopg2-binary==2.9.10
